Fix some animation bugs & some other changes

+++ app/src/main/java/com/genonbeta/TrebleShot/object/TransferObject.java
@@ -45,7 +45,7 @@ public class TransferObject
 		this.type = type;
 	}
 
-	public TransferObject(int requestId)
+	public TransferObject(long requestId)
 	{
 		this.requestId = requestId;
 	}
